+++ Sources/KeyboardKit/Gestures/GestureButton/GestureButtonTimer.swift
@@ -17,7 +17,7 @@ public class GestureButtonTimer: ObservableObject {
     /// - Parameters:
     ///   - interval: The trigger interval, by default `0.1`.
     public init(
-        interval: TimeInterval = 0.4
+        interval: TimeInterval = 0.1
     ) {
         self.interval = interval
     }